# src/Server/IoTGateway/modules/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.5",
    "title": "REST API Test Docs",
    "description": "該文件可提供在該專案中所需提供的功能"
  },
  "host": "localhost:3095",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/upload/{deviceID}/data": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "hum": {
                  "example": "any"
                },
                "temp": {
                  "example": "any"
                },
                "tvoc": {
                  "example": "any"
                },
                "co": {
                  "example": "any"
                },
                "co2": {
                  "example": "any"
                },
                "pm25": {
                  "example": "any"
                },
                "o3": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/StatusGet/{deviceID}/powerStatus": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/read/{deviceID}/ALL": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/hum": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/temp": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/tvoc": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/co2": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/co": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/pm25": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/{deviceID}/o3": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}